Translation updates

+++ translation_converter.py
@@ -7,45 +7,53 @@ import sys
 import polib
 import subprocess
 
-subprocess.call(["lupdate", "pamac_tray.cpp", "-ts", "pamac_tray.ts"])
+if len(sys.argv) == 1:
+    print(f'Usage: {sys.argv[0]} <path-to-po-files>')
+    exit(0)
+
+subprocess.call(["lupdate", "pamac_tray.cpp", "main.cpp", "-ts", "pamac_tray.ts"])
 subprocess.call(["lconvert", "pamac_tray.ts", "-o", "pamac_tray.po"])
 os.remove('pamac_tray.ts')
 
-sourcefile = polib.pofile("pamac_tray.po")
-print(sourcefile.metadata)
+tray_icon_source_po = polib.pofile("pamac_tray.po")
 
-for filename in os.listdir(sys.argv[1]):
-    if not filename.endswith(".po"):
+for pamac_po_filename in os.listdir(sys.argv[1]):
+
+    if not pamac_po_filename.endswith(".po"):
         continue
-    print(filename)
-    base = os.path.basename(filename)
+
+    base = os.path.basename(pamac_po_filename)
     language = os.path.splitext(base)[0]
 
-    file = polib.pofile(sys.argv[1]+filename)
+    pamac_po = polib.pofile(sys.argv[1] + "/" + pamac_po_filename)
 
     result = polib.POFile()
-    result.metadata = sourcefile.metadata
+    result.metadata = tray_icon_source_po.metadata
 
-    lst = file.translated_entries()
+    pamac_messages = pamac_po.translated_entries()
 
-    for el in lst:
-        e = next((e for e in sourcefile if e.msgid == el.msgid.replace('%u', '%n')), None)
-        if e and ('src/tray.vala', '') in el.occurrences:
+    for pamac_message in pamac_messages:
+        if not ('src/tray.vala', '') in pamac_message.occurrences:
+            continue
 
-            if not el.msgstr_plural:
-                e.msgstr = el.msgstr
+        tray_messages = [e for e in tray_icon_source_po if e.msgid == pamac_message.msgid.replace('%u', '%n').replace('_', '')]
+        if not tray_messages:
+            continue
+
+        for tray_message in tray_messages:
+
+            if not pamac_message.msgstr_plural:
+                tray_message.msgstr = pamac_message.msgstr.replace('_', '')
             else:
-                e.msgstr_plural = el.msgstr_plural
-                for plural in e.msgstr_plural:
-                    e.msgstr_plural[plural] = e.msgstr_plural[plural].replace('%u', '%n')
+                tray_message.msgstr_plural = pamac_message.msgstr_plural
+                for plural in tray_message.msgstr_plural:
+                    tray_message.msgstr_plural[plural] = tray_message.msgstr_plural[plural].replace('%u', '%n')
 
-            result.append(e)
+            result.append(tray_message)
 
     result.save(f"pamac_tray_{language}.po")
 
     subprocess.call(["lconvert", f"pamac_tray_{language}.po", "-o", f"pamac_tray_{language}.ts"])
-
-
     subprocess.call(["lrelease", f"pamac_tray_{language}.ts"])
     os.remove(f"pamac_tray_{language}.ts")
     os.remove(f"pamac_tray_{language}.po")
